requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pyzmq==20.0.0
qlua==1.3.0
six==1.15.0
toml==0.10.2
